requirements.txt
pinecone-client>=3.0.0
sentence-transformers>=2.6.0
pypdf>=4.0.0
tqdm>=4.66.0
torch>=2.0.0
transformers>=4.38.0